discord-bot/requirements.txt
discord.py==2.2.2
prisma==0.8.2
prisma_client==0.2.1
python-dotenv==1.0.0
pytz==2022.1